webnfc/manual-test/index.html

<meta charset=utf-8>
<title>Web NFC Watch Tests</title>
<link rel="author" title="Intel" href="http://www.intel.com"/>
<link rel="author" title="Zhao Ming" href="mailto:mingx.zhao@intel.com"/>
<link rel="help" href="https://w3c.github.io/web-nfc/#the-nfc-interface"/>
<script src="./resources/testharness.js"></script>
<script src="./resources/testharnessreport.js"></script>
<script src="./nfc_help.js"></script>

<div id="log"></div>
<script>
//  navigator.nfc.watch((message) => {
//    if (message.data[0].recordType == 'empty') {
//      console.log("DEBUG-1")
//      navigator.nfc.push({
//        url: "/custom/path",
//        data: [{ recordType: "text", data: 'Hello World' }]
//      });
//    } else {
//      console.log("DEBUG-2")
//      console.log('Read message written by ' + message.url);
//      processMessage(message);
//    }
//  }).then(() => {
//    console.log("DEBUG-3")
//    console.log("Added a watch.");
//  }).catch((error) => {
//    console.log("DEBUG-4")
//    console.log("Adding watch failed: " + error.name);
//  });
//  
//  function processMessage(message) {
//    console.log("DEBUG-5")
//    console.log('-----------------------------')
//    for (let record of message.data) {
//      console.log("DEBUG-6")
//      switch (record.recordType) {
//        case "text":
//          console.log("DEBUG-7")
//          console.log('Data is text: ' + record.data);
//          break;
//        case "url":
//          console.log('Data is URL: ' + record.data);
//          break;
//        case "json":
//          console.log('JSON data: ' + record.data.myProperty.toString());
//          break;
//        case "opaque":
//          if (record.mediaType == 'image/png') {
//            var img = document.createElement("img");
//            img.src = URL.createObjectURL(new Blob(record.data, record.mediaType));
//            img.onload = function() {
//              window.URL.revokeObjectURL(this.src);
//            };
//          }
//          break;
//      }
//    }
//  }

//function pushCheck(arg1, arg2, arg3) {
  // arg1 : context    arg2 : type     arg3 : comment 
//  promise_test(t => {
//    return navigator.nfc.push("aaabbbccc")
//      .then(() => {
//        navigator.nfc.watch((message)
//            .then(()=> {
//              console.log(message.data.recordType)
//              assert_equals(message.data.recordType, "text");
//            })
//        )})
//  }, "test comm")
//}
let a = "test strings zhaoming"
let b = "1text"
let c = "this is test case for automachie"
//pushCheck(a, b , c)
promise_test(t => {
  return navigator.nfc.push("aaaaaa")
    .then(() => {
      navigator.nfc.watch((message) => {
          console.log(message.data.recordType);
          assert_unreached();
      }).then(()=>{
          assert_unreached("22222222222222")

        
      })
    })
  })


// promise_test(t=> {
//   return navigator.nfc.push("12345678")
//   return navigator.nfc.cancelPush()
// }, "1111123456789")
//promise_test(t => {
//  new Promise((resolve,reject) => {
//    return navigator.nfc.push("00000000000").then((data) => {
//      console.log(data);
//      console.log("---------")
//      return navigator.nfc.cancelPush();
//      resolve();
//    })
////  })
////})
////promise_test(t=> {
////  return  promise_rejects(t, "AbortError", navigator.nfc.push("aaa"))
////    navigator.nfc.cancelPush()
////    console.log("1111111111111111111111111111111111111111111111111111111111111111111111111")
////})
//
//promise_test(t=> {
//  let promise = navigator.nfc.push(test_text_data);
//  navigator.nfc.cancelPush();
//  return promise_rejects(t, 'AbortError', promise);
//}, "nfc.cancelPush should reject pending promise with AbortError");
</script>
